dev: extract worker helpers
